attempt to fix linter
